Init replace error returns with exceptions
