+++ openapi_python_client/config.py
@@ -24,9 +24,9 @@ class Config(BaseModel):
     """
 
     class_overrides: Dict[str, ClassOverride] = {}
-    project_name_override: Optional[str]
-    package_name_override: Optional[str]
-    package_version_override: Optional[str]
+    project_name_override: Optional[str] = None
+    package_name_override: Optional[str] = None
+    package_version_override: Optional[str] = None
     use_path_prefixes_for_title_model_names: bool = True
     post_hooks: List[str] = [
         "autoflake -i -r --remove-all-unused-imports --remove-unused-variables --ignore-init-module-imports .",
