+++ openapi_python_client/parser/__init__.py
@@ -2,4 +2,5 @@
 
 __all__ = ["GeneratorData", "import_string_from_class"]
 
-from .openapi import GeneratorData, import_string_from_class
+from .openapi import GeneratorData
+from .endpoints import import_string_from_class
