+++ openapi_python_client/parser/properties/property.py
@@ -158,7 +158,9 @@ class Property:
     @property
     def is_base_type(self) -> bool:
         """Base types, represented by any other of `Property` than `ModelProperty` should not be quoted."""
-        from . import ListProperty, ModelProperty, UnionProperty
+        from .list_property import ListProperty
+        from .model_property import ModelProperty
+        from .union_property import UnionProperty
 
         return self.__class__.__name__ not in {
             ModelProperty.__name__,
