+++ openapi_python_client/templates/endpoint_macros.py.jinja
@@ -161,7 +161,7 @@ Args:
 
 {% endif %}
 Raises:
-    errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
+    errors.UnexpectedStatus: If the server returns an undocumented status code.
     httpx.TimeoutException: If the request takes longer than Client.timeout.
 
 Returns:
