+++ openapi_python_client/templates/errors.py.jinja
@@ -1,7 +1,7 @@
 """ Contains shared errors types that can be raised from API functions """
 
 class UnexpectedStatus(Exception):
-    """ Raised by api functions when the response status an undocumented status and Client.raise_on_unexpected_status is True """
+    """ Raised by api functions when the response status an undocumented status. """
 
     def __init__(self, status_code: int, content: bytes):
         self.status_code = status_code
