+++ openapi_python_client/templates/security.py.jinja
@@ -1,0 +1,24 @@
+from typing import Literal
+
+from dlt.common.configuration.specs.base_configuration import BaseConfiguration, configspec
+
+from openapi_python_client.security_specs import ApiKeyCredentialsBase, HttpBasicCredentialsBase, HttpBearerCredentialsBase, OAuth2CredentialsBase
+
+
+{% for scheme in security_schemes %}
+{% if scheme.scheme.type == "apiKey" %}
+class {{ scheme.class_name }}(ApiKeyCredentialsBase):
+    location: Literal["{{ scheme.scheme.security_scheme_in }}"] = "{{ scheme.scheme.security_scheme_in }}"
+    name: Literal["{{ scheme.name }}"] = "{{ scheme.name }}"
+    
+{% elif scheme.type == "http" and scheme.scheme == "basic" %}
+class {{ scheme.class_name }}(HttpBasicCredentialsBase):
+    pass
+{% elif scheme.type == "http" and scheme.scheme == "bearer" %}
+class {{ scheme.class_name }}(HttpBearerCredentialsBase):
+    pass
+{% elif scheme.type == "oauth2" %}
+class {{ scheme.class_name }}(OAuth2CredentialsBase):
+    pass
+{% endif %}
+{% endfor %}
