+++ openapi_python_client/templates/security_scheme.py.jinja
@@ -1,0 +1,22 @@
+from typing import Literal
+
+from dlt.common.configuration.specs.base_configuration import BaseConfiguration, configspec
+
+from openapi_python_client.security_specs import ApiKeyCredentialsBase, HttpBasicCredentialsBase, HttpBearerCredentialsBase, OAuth2CredentialsBase
+
+
+{% if scheme.data.type == "apiKey" %}
+class {{ scheme.class_name }}(ApiKeyCredentialsBase):
+    location: Literal["{{ scheme.data.security_scheme_in }}"] = "{{ scheme.data.security_scheme_in }}"
+    name: Literal["{{ scheme.name }}"] = "{{ scheme.name }}"
+    
+{% elif scheme.type == "http" and scheme.data == "basic" %}
+class {{ scheme.class_name }}(HttpBasicCredentialsBase):
+    pass
+{% elif scheme.type == "http" and scheme.data == "bearer" %}
+class {{ scheme.class_name }}(HttpBearerCredentialsBase):
+    pass
+{% elif scheme.type == "oauth2" %}
+class {{ scheme.class_name }}(OAuth2CredentialsBase):
+    pass
+{% endif %}
