+++ openapi_python_client/templates/types.py.jinja
@@ -39,7 +39,7 @@ class Response(Generic[T]):
     status_code: HTTPStatus
     content: bytes
     headers: MutableMapping[str, str]
-    parsed: Optional[T]
+    parsed: T
 
 
 __all__ = ["File", "Response", "FileJsonType"]
